+++ Makefile.libcoap
@@ -2,4 +2,11 @@ LIBCOAP_DIR = os/net/app-layer/libcoap
 MODULES += $(LIBCOAP_DIR)/src
 MODULES += $(LIBCOAP_DIR)/include
 MODULES_SOURCES_EXCLUDES += coap_io_riot.c
+MODULES_SOURCES_EXCLUDES += coap_strm_riot.c
+MODULES_SOURCES_EXCLUDES += coap_dgrm_riot.c
 MODULES_SOURCES_EXCLUDES += coap_io_lwip.c
+MODULES_SOURCES_EXCLUDES += coap_strm_lwip.c
+MODULES_SOURCES_EXCLUDES += coap_dgrm_lwip.c
+MODULES_SOURCES_EXCLUDES += coap_io_posix.c
+MODULES_SOURCES_EXCLUDES += coap_strm_posix.c
+MODULES_SOURCES_EXCLUDES += coap_dgrm_posix.c
