+++ examples/riot/pkg_libcoap/Makefile.libcoap
@@ -6,6 +6,7 @@ SRC := coap_address.c \
   coap_block.c \
   coap_cache.c \
   coap_debug.c \
+  coap_dgrm_riot.c \
   coap_dtls.c \
   coap_encode.c \
   coap_event.c \
@@ -29,8 +30,8 @@ SRC := coap_address.c \
   coap_session.c \
   coap_sha1.c \
   coap_str.c \
+  coap_strm_riot.c \
   coap_subscribe.c \
-  coap_tcp.c \
   coap_threadsafe.c \
   coap_tinydtls.c \
   coap_uri.c \
