+++ src/coap_layers.c
@@ -45,41 +45,41 @@ coap_layer_func_t coap_layers_coap[COAP_PROTO_LAST][COAP_LAYER_LAST] = {
   },
   {
     /* COAP_PROTO_UDP */
-    { NULL,                 coap_netif_dgrm_write, coap_session_establish, coap_netif_close }, /* SESSION */
-    { NULL,                 NULL,                  NULL,                   NULL             }, /* WS */
-    { NULL,                 NULL,                  NULL,                   NULL             }  /* TLS */
+    { NULL,                 coap_netif_dgrm_write, coap_session_establish, coap_netif_dgrm_close }, /* SESSION */
+    { NULL,                 NULL,                  NULL,                   NULL                  }, /* WS */
+    { NULL,                 NULL,                  NULL,                   NULL                  }  /* TLS */
   },
   {
     /* COAP_PROTO_DTLS */
-    { NULL,                 coap_dtls_send,        coap_dtls_establish,    coap_dtls_close  }, /* SESSION */
-    { NULL,                 NULL,                  NULL,                   NULL             }, /* WS */
-    { NULL,                 coap_netif_dgrm_write, coap_session_establish, coap_netif_close }  /* TLS */
+    { NULL,                 coap_dtls_send,        coap_dtls_establish,    coap_dtls_close       }, /* SESSION */
+    { NULL,                 NULL,                  NULL,                   NULL                  }, /* WS */
+    { NULL,                 coap_netif_dgrm_write, coap_session_establish, coap_netif_dgrm_close }  /* TLS */
   },
 #if !COAP_DISABLE_TCP
   {
     /* COAP_PROTO_TCP */
-    { coap_netif_strm_read, coap_netif_strm_write, coap_session_establish, coap_netif_close }, /* SESSION */
-    { NULL,                 NULL,                  NULL,                   NULL             }, /* WS */
-    { NULL,                 NULL,                  NULL,                   NULL             }  /* TLS */
+    { coap_netif_strm_read, coap_netif_strm_write, coap_session_establish, coap_netif_strm_close }, /* SESSION */
+    { NULL,                 NULL,                  NULL,                   NULL                  }, /* WS */
+    { NULL,                 NULL,                  NULL,                   NULL                  }  /* TLS */
   },
   {
     /* COAP_PROTO_TLS */
-    { coap_tls_read,        coap_tls_write,        coap_tls_establish,     coap_tls_close   }, /* SESSION */
-    { NULL,                 NULL,                  NULL,                   NULL             }, /* WS */
-    { coap_netif_strm_read, coap_netif_strm_write, coap_session_establish, coap_netif_close }  /* TLS */
+    { coap_tls_read,        coap_tls_write,        coap_tls_establish,     coap_tls_close        }, /* SESSION */
+    { NULL,                 NULL,                  NULL,                   NULL                  }, /* WS */
+    { coap_netif_strm_read, coap_netif_strm_write, coap_session_establish, coap_netif_strm_close }  /* TLS */
   },
 #if COAP_WS_SUPPORT
   {
     /* COAP_PROTO_WS */
-    { coap_ws_read,         coap_ws_write,         coap_ws_establish,      coap_ws_close    }, /* SESSION */
-    { coap_netif_strm_read, coap_netif_strm_write, coap_session_establish, coap_netif_close }, /* WS */
-    { NULL,                 NULL,                  NULL,                   NULL             }  /* TLS */
+    { coap_ws_read,         coap_ws_write,         coap_ws_establish,      coap_ws_close         }, /* SESSION */
+    { coap_netif_strm_read, coap_netif_strm_write, coap_session_establish, coap_netif_strm_close }, /* WS */
+    { NULL,                 NULL,                  NULL,                   NULL                  }  /* TLS */
   },
   {
     /* COAP_PROTO_WSS */
-    { coap_ws_read,         coap_ws_write,         coap_tls_establish,     coap_ws_close    }, /* SESSION */
-    { coap_tls_read,        coap_tls_write,        coap_session_establish, coap_tls_close   }, /* WS */
-    { coap_netif_strm_read, coap_netif_strm_write, coap_ws_establish,      coap_netif_close }  /* TLS */
+    { coap_ws_read,         coap_ws_write,         coap_tls_establish,     coap_ws_close         }, /* SESSION */
+    { coap_tls_read,        coap_tls_write,        coap_session_establish, coap_tls_close        }, /* WS */
+    { coap_netif_strm_read, coap_netif_strm_write, coap_ws_establish,      coap_netif_strm_close }  /* TLS */
   }
 #else /* !COAP_WS_SUPPORT */
   {
