+++ win32/libcoap.vcxproj
@@ -44,12 +44,14 @@
     <ClCompile Include="..\src\coap_block.c" />
     <ClCompile Include="..\src\coap_cache.c" />
     <ClCompile Include="..\src\coap_debug.c" />
+    <ClCompile Include="..\src\coap_dgrm_posix.c" />
     <ClCompile Include="..\src\coap_dtls.c" />
     <ClCompile Include="..\src\coap_encode.c" />
     <ClCompile Include="..\src\coap_event.c" />
     <ClCompile Include="..\src\coap_hashkey.c" />
     <ClCompile Include="..\src\coap_gnutls.c" />
     <ClCompile Include="..\src\coap_io.c" />
+    <ClCompile Include="..\src\coap_io_posix.c" />
     <ClCompile Include="..\src\coap_layers.c" />
     <ClCompile Include="..\src\coap_mbedtls.c" />
     <ClCompile Include="..\src\coap_mem.c" />
@@ -66,9 +68,9 @@
     <ClCompile Include="..\src\coap_session.c" />
     <ClCompile Include="..\src\coap_sha1.c" />
     <ClCompile Include="..\src\coap_str.c" />
+    <ClCompile Include="..\src\coap_strm_posix.c" />
     <ClCompile Include="..\src\coap_subscribe.c" />
     <ClCompile Include="..\src\coap_time.c" />
-    <ClCompile Include="..\src\coap_tcp.c" />
     <ClCompile Include="..\src\coap_threadsafe.c" />
     <ClCompile Include="..\src\coap_tinydtls.c" />
     <ClCompile Include="..\src\coap_uri.c" />
